Added a significant amount of content
Is this a good log message?  How could I have done better here?

I've added slides that take us up through our first commit of a README
file.  Much of the presentation is still in template mode.  I am still
working in shotgun mode adding new slides and refining previous
slides.  It is not possible to state all of the changes I have made in
this log. We can debate whether this strategy is appropraite or not.
I find that during the initial content building phase it is too
tedious to describe the updates, say slide-by slide.  Instead, in the
beginning, I approach it using drafts.  For example I am still working
on the initial rough draft of the presentation.  Once we're through
the draft phase we should be much more specific about the changes
being made.  For code it is differnet, one can more easily commit
function-by-function.  The shotgun mode doesn't work as well with
code.

Note how long and descriptive this commiut log message is.  Of course
this is overkill but it goes to my point that detailed records can be
maintained easily using git.

Also note how the first line is treated like a header.  The first line
should be short and descriptive followed by a few line breaks.  More
detailed notes are written beneath this header.  This is done for
display reasons.