Edited "Incorporate into Repository" section
To be honest it is a little difficult to describe changes to slides in
the commit log.  With code it is more natural.  But until the
revisions to the presentation become small it is hard to be more
specific than "Edited such and such."  It would be crazy to list each
typo that was fixed, or grammar that was clarified.  Even if it is
less useful than commit logs associated with code we must still try to
maintain a useful record of the changes to the slides.